+++ easy_thumbnails/engine.py
@@ -55,12 +55,17 @@ def save_image(image, destination=None, filename=None, **options):
         options.setdefault('quality', 85)
     saved = False
     if format == 'JPEG':
+        if settings.THUMBNAIL_PROGRESSIVE and (
+                max(image.size) >= settings.THUMBNAIL_PROGRESSIVE):
+            options['progressive'] = True
         try:
             image.save(destination, format=format, optimize=1, **options)
             saved = True
         except IOError:
             # Try again, without optimization (PIL can't optimize an image
-            # larger than ImageFile.MAXBLOCK, which is 64k by default)
+            # larger than ImageFile.MAXBLOCK, which is 64k by default). This
+            # shouldn't be triggered very often these days, as recent versions
+            # of pillow avoid the MAXBLOCK limitation.
             pass
     if not saved:
         image.save(destination, format=format, **options)